CLI implemented

+++ src/main/scala/Speedrun.scala
@@ -3,13 +3,14 @@ package project0
 import play.api.libs.json._
 import org.bson.types.ObjectId
 
+
 //TODO: Add speedrunID and possibly other fields as well
 case class Speedrun(
     _id : ObjectId,
     place: Int, //place in current leaderboard
     weblink: String, //link to speedrun's page on speedrun.com
     date: String,
-    time: Int   //total time of speedrun (in seconds)
+    time: Float   //total time of speedrun (in seconds)
 ){
 
     override def toString: String = {
@@ -18,6 +19,6 @@ case class Speedrun(
 }
 
 object Speedrun{
-  def apply(place: Int, weblink: String, date: String, time: Int): Speedrun =
+  def apply(place: Int, weblink: String, date: String, time: Float): Speedrun =
     Speedrun(new ObjectId(), place, weblink, date, time)
 }
